ecommerce/auth/requirements.txt
httpx
fastapi
uvicorn
sqlalchemy
pydantic
requests
bcrypt
env
PyMySQL
websockets
socketio
pydantic_core
cryptography
python-jose
pytz
passlib
pytest
python-dotenv
httpx